relates to #1983: moved metrics config to common project
