+++ apis/sgv2-quarkus-common/src/test/java/io/stargate/sgv2/api/common/metrics/TenantRequestMetricsFilterTest.java
@@ -14,52 +14,74 @@
  * limitations under the License.
  *
  */
-
-package io.stargate.sgv2.docsapi.api.common.metrics.configuration;
+package io.stargate.sgv2.api.common.metrics;
 
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.google.common.collect.ImmutableMap;
 import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.QuarkusTestProfile;
 import io.quarkus.test.junit.TestProfile;
+import io.stargate.sgv2.api.common.testprofiles.FixedTenantTestProfile;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
 import org.junit.jupiter.api.Test;
 
 @QuarkusTest
-@TestProfile(MicrometerConfigurationTest.Profile.class)
-class MicrometerConfigurationTest {
+@TestProfile(TenantRequestMetricsFilterTest.Profile.class)
+class TenantRequestMetricsFilterTest {
 
-  public static class Profile implements QuarkusTestProfile {
+  public static class Profile extends FixedTenantTestProfile {
 
     @Override
     public Map<String, String> getConfigOverrides() {
+      Map<String, String> configOverrides = super.getConfigOverrides();
+
       return ImmutableMap.<String, String>builder()
-          .put("stargate.metrics.global-tags.module", "test-module")
-          .put("stargate.metrics.global-tags.extra-tag", "extra-value")
+          .putAll(configOverrides)
+          .put("quarkus.micrometer.export.prometheus.path", "/metrics")
+          .put("stargate.metrics.tenant-request-counter.metric-name", "test.metrics")
+          .put("stargate.metrics.tenant-request-counter.tenant-tag", "tenantTag")
+          .put("stargate.metrics.tenant-request-counter.error-tag", "errorTag")
           .build();
     }
   }
 
+  // simple testing endpoint, helps in testing metrics
+  @Path("testing")
+  public static class TestingResource {
+
+    @GET
+    @Produces("text/plain")
+    public String greet() {
+      return "hello";
+    }
+  }
+
   @Test
-  public void globalTags() {
+  public void record() {
+    // call endpoint
+    given().when().get("/testing").then().statusCode(200);
+
     // collect metrics
     String result = given().when().get("/metrics").then().statusCode(200).extract().asString();
 
+    // find target metrics
     List<String> meteredLines =
         Arrays.stream(result.split(System.getProperty("line.separator")))
+            .filter(line -> line.startsWith("test_metrics_total"))
             .collect(Collectors.toList());
 
-    // confirm existence in one, not all support global tags
     assertThat(meteredLines)
         .anySatisfy(
             metric ->
                 assertThat(metric)
-                    .contains("module=\"test-module\"")
-                    .contains("extra_tag=\"extra-value\""));
+                    .contains("tenantTag=\"" + FixedTenantTestProfile.TENANT_ID + "\"")
+                    .contains("errorTag=\"false\""));
   }
 }
